requirements.txt
openai==2.7.2
python-dotenv==1.0.0